Refactor: NULS is no longer used by Aleph.
Solution: Remove the possibility to create NULS Account with the SDK.
