+++ src/accounts/index.ts
@@ -4,6 +4,5 @@ import * as ethereum from "./ethereum";
 import * as solana from "./solana";
 import * as substrate from "./substrate";
 import * as nuls2 from "./nuls2";
-import * as nuls from "./nuls";
 
-export { avalanche, base, ethereum, solana, substrate, nuls, nuls2 };
+export { avalanche, base, ethereum, solana, substrate, nuls2 };
