+++ tests/index.ts
@@ -1,7 +1,6 @@
 import * as ethereum from "../src/accounts/ethereum";
 import * as solana from "../src/accounts/solana";
 import * as nuls2 from "../src/accounts/nuls2";
-import * as nuls from "../src/accounts/nuls";
 import * as avalanche from "../src/accounts/avalanche";
 
 import * as aggregate from "../src/messages/aggregate/index";
@@ -10,4 +9,4 @@ import * as store from "../src/messages/store/index";
 import * as forget from "../src/messages/forget/index";
 import * as program from "../src/messages/program/index";
 
-export { avalanche, ethereum, solana, nuls, nuls2, aggregate, post, store, forget, program };
+export { avalanche, ethereum, solana, nuls2, aggregate, post, store, forget, program };
